move scroll restoration logic into its own hook
