+++ packages/lightnet/src/pages/search-page/components/SearchList.tsx
@@ -7,6 +7,7 @@ import SearchListItem, {
   type MediaType,
   type TranslatedLanguage,
 } from "./SearchListItem"
+import { useRestoreScrollPosition } from "../hooks/use-restore-position"
 
 interface Props {
   currentLocale: string | undefined
@@ -27,7 +28,6 @@ export default function SearchList({
   showLanguage,
   mediaTypes,
 }: Props) {
-  const restoredPosition = useRef(0)
   const listRef = useRef<HTMLDivElement | null>(null)
   const [rowHeight, setRowHeight] = useState(208)
 
@@ -62,43 +62,7 @@ export default function SearchList({
     }
   }, [])
 
-  // Astro's ClientRouter is writing a incorrect scroll position to
-  // history.state.scrollY. We have seen this when having the search results
-  // below the fold. This leads to incorrect scroll position restoration after
-  // navigating back.
-  //
-  // We store the correct scroll value to a custom property "searchScrollY"
-  // to be used by the scroll restoration.
-  useEffect(() => {
-    restoredPosition.current = history.state?.searchScrollY ?? 0
-    let timeout: number | undefined
-    const storeScrollPosition = () => {
-      if (timeout) {
-        window.clearTimeout(timeout)
-      }
-      timeout = window.setTimeout(() => {
-        const state = history.state ?? {}
-        history.pushState({ ...state, searchScrollY: window.scrollY }, "")
-      }, 50)
-    }
-    document.addEventListener("scroll", storeScrollPosition)
-    return () => {
-      document.removeEventListener("scroll", storeScrollPosition)
-    }
-  }, [])
-
-  // Restore scroll position after back navigation.
-  // We need to implement this because default restoration is kicking in
-  // before results are loaded. Also default restoration uses and incorrect scroll position.
-  // This is why we rely on our custom property.
-  useEffect(() => {
-    const { state } = history
-    if (!isLoading && state?.searchScrollY !== undefined) {
-      requestAnimationFrame(() =>
-        window.scrollTo({ top: restoredPosition.current }),
-      )
-    }
-  }, [isLoading])
+  useRestoreScrollPosition(isLoading)
 
   const t = (key: TranslationKey) => translations[key]
